requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
pymongo==4.6.1
dnspython==2.3.0
python-dotenv==1.0.1
pyjwt==2.8.0
Werkzeug==2.3.7
itsdangerous==2.1.2
Flask-Cors==4.0.0
Pillow==10.2.0
